Introduce DataFrameLayer
