+++ dask/dataframe/io/tests/test_parquet.py
@@ -17,7 +17,7 @@ from dask.blockwise import Blockwise, optimize_blockwise
 from dask.dataframe._compat import PANDAS_GT_110, PANDAS_GT_121
 from dask.dataframe.utils import assert_eq
 from dask.dataframe.io.parquet.utils import _parse_pandas_metadata
-from dask.dataframe.optimize import optimize_blockwise_columnar_getitem
+from dask.dataframe.optimize import optimize_dataframe_getitem
 from dask.dataframe.io.parquet.core import BlockwiseParquet
 from dask.utils import natural_sort_key, parse_bytes
 
@@ -2403,7 +2403,7 @@ def test_getitem_optimization(tmpdir, engine, preserve_index, index):
     # Write ddf back to disk to check that the round trip
     # preserves the getitem optimization
     out = ddf.to_frame().to_parquet(tmp_path_wt, engine=engine, compute=False)
-    dsk = optimize_blockwise_columnar_getitem(out.dask, keys=[out.key])
+    dsk = optimize_dataframe_getitem(out.dask, keys=[out.key])
 
     read = [key for key in dsk.layers if key.startswith("read-parquet")][0]
     subgraph = dsk.layers[read]
@@ -2420,7 +2420,7 @@ def test_getitem_optimization_empty(tmpdir, engine):
     ddf.to_parquet(fn, engine=engine)
 
     df2 = dd.read_parquet(fn, columns=[], engine=engine)
-    dsk = optimize_blockwise_columnar_getitem(df2.dask, keys=[df2._name])
+    dsk = optimize_dataframe_getitem(df2.dask, keys=[df2._name])
 
     subgraph = next(iter((dsk.layers.values())))
     assert isinstance(subgraph, BlockwiseParquet)
@@ -2757,7 +2757,7 @@ def test_read_parquet_getitem_skip_when_getting_read_parquet(tmpdir, engine):
 
     # Make sure we are still allowing the getitem optimization
     ddf = ddf["A"]
-    dsk = optimize_blockwise_columnar_getitem(ddf.dask, keys=[(ddf._name, 0)])
+    dsk = optimize_dataframe_getitem(ddf.dask, keys=[(ddf._name, 0)])
     read = [key for key in dsk.layers if key.startswith("read-parquet")][0]
     subgraph = dsk.layers[read]
     assert isinstance(subgraph, BlockwiseParquet)
